+++ lib/generators/cms/kickstart/templates/app/models/website.rb
@@ -3,6 +3,10 @@ class Website < Obj
     self.class.default_homepage
   end
 
+  def homepages
+    @homepages ||= toclist.select { |obj| obj.is_a?(Homepage) }
+  end
+
   def website
     self
   end